// src/domains/service-clients/course/proto/generated/course/types/course.ts
export interface CourseData {
  id: string;
  title: string;
  topics: string[];
  instructorId: string;
  subTitle?: string | undefined;
  category: string;
  subCategory?: string | undefined;
  language?: string | undefined;
  subtitleLanguage?: string | undefined;
  level?: string | undefined;
  durationValue?: string | undefined;
  durationUnit?: string | undefined;
  description?: string | undefined;
  learningOutcomes: string[];
  targetAudience: string[];
  requirements: string[];
  thumbnail?: string | undefined;
  trailer?: string | undefined;
  status?: string | undefined;
  slug?: string | undefined;
  rating?: number | undefined;
  numberOfRating?: number | undefined;
  students?: number | undefined;
  sections: SectionData[];
  createdAt: string;
  updatedAt: string;
  deletedAt?: string | undefined;
  price?: number | undefined;
  discountPrice?: number | undefined;
  currency?: string | undefined;
  instructor?: User | undefined;
}

export interface CourseMetadata {
  id: string;
  title: string;
  topics: string[];
  instructorId: string;
  subTitle: string;
  category: string;
  subCategory: string;
  language: string;
  subtitleLanguage: string;
  level: string;
  durationValue: string;
  durationUnit: string;
  description?: string | undefined;
  learningOutcomes: string[];
  targetAudience: string[];
  requirements: string[];
  thumbnail?: string | undefined;
  trailer?: string | undefined;
  status: string;
  slug: string;
  rating: number;
  numberOfRating: number;
  students: number;
  createdAt: string;
  updatedAt: string;
  deletedAt?: string | undefined;
  price?: number | undefined;
  noOfLessons: number;
  noOfSections: number;
  noOfQuizzes: number;
  discountPrice?: number | undefined;
  currency?: string | undefined;
  instructor: User | undefined;
}