// src/domains/service-clients/course/proto/generated/course/types/lesson.ts
export interface ContentMetaData {
  title?: string | undefined;
  fileName?: string | undefined;
  mimeType?: string | undefined;
  fileSize?: string | undefined;
  url?: string | undefined;
}

export interface DeleteLessonRequest {
  lessonId: string;
  courseId: string;
  userId: string;
}

export interface GetLessonsBySectionRequest {
  sectionId: string;
}

export interface LessonData {
  id: string;
  sectionId: string;
  isPreview: boolean;
  description: string;
  estimatedDuration: number;
  order: number;
  title: string;
  isPublished: boolean;
  contentType: string;
  contentUrl: string;
  metadata: ContentMetaData | undefined;
  createdAt: string;
  updatedAt: string;
  deletedAt?: string | undefined;
}
// ...
export const ContentMetaData: MessageFns<ContentMetaData> = {
  encode(message: ContentMetaData, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.title !== undefined) {
      writer.uint32(10).string(message.title);
    }
    if (message.fileName !== undefined) {
      writer.uint32(18).string(message.fileName);
    }
    if (message.mimeType !== undefined) {
      writer.uint32(26).string(message.mimeType);
    }
    if (message.fileSize !== undefined) {
      writer.uint32(34).string(message.fileSize);
    }
    if (message.url !== undefined) {
      writer.uint32(42).string(message.url);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ContentMetaData {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseContentMetaData();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.title = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.fileName = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.mimeType = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.fileSize = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.url = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ContentMetaData {
    return {
      title: isSet(object.title) ? globalThis.String(object.title) : undefined,
      fileName: isSet(object.fileName)
        ? globalThis.String(object.fileName)
        : isSet(object.file_name)
        ? globalThis.String(object.file_name)
        : undefined,
      mimeType: isSet(object.mimeType)
        ? globalThis.String(object.mimeType)
        : isSet(object.mime_type)
        ? globalThis.String(object.mime_type)
        : undefined,
      fileSize: isSet(object.fileSize)
        ? globalThis.String(object.fileSize)
        : isSet(object.file_size)
        ? globalThis.String(object.file_size)
        : undefined,
      url: isSet(object.url) ? globalThis.String(object.url) : undefined,
    };
  },

  toJSON(message: ContentMetaData): unknown {
    const obj: any = {};
    if (message.title !== undefined) {
      obj.title = message.title;
    }
    if (message.fileName !== undefined) {
      obj.fileName = message.fileName;
    }
    if (message.mimeType !== undefined) {
      obj.mimeType = message.mimeType;
    }
    if (message.fileSize !== undefined) {
      obj.fileSize = message.fileSize;
    }
    if (message.url !== undefined) {
      obj.url = message.url;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ContentMetaData>, I>>(base?: I): ContentMetaData {
    return ContentMetaData.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ContentMetaData>, I>>(object: I): ContentMetaData {
    const message = createBaseContentMetaData();
    message.title = object.title ?? undefined;
    message.fileName = object.fileName ?? undefined;
    message.mimeType = object.mimeType ?? undefined;
    message.fileSize = object.fileSize ?? undefined;
    message.url = object.url ?? undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}